Don't push IME above all app windows and dock divider.
The IME should only be pushed above the application that requested it.
There might be other windows that go above the IME, e.g. spinner dialog.

We used to need the IME to go above all app windows, because the dock
divider goes above all app windows and we want IME above the dock
divider. This is not necessary anymore, because the dock divider is now
aware of the IME and changes it's size to available space. The
transition is also graceful enough, that it looks correct even when the
IME goes away.

Bug: 26325071
Change-Id: I8c05f2d200c5bb92c5fdb0d252e277fbe43a60a5