Fix black screen when launching app from Recents into non-docked area
In this case, we're putting the app into a fullscreen stack that is
resized due to docked stack. When setting up the thumbnail scale up
animation, we have to treat this case as animating into freeform window,
as the containing frame will not be fullscreen, and it's not a clip
rect. For freeform windows, we set up scaling and translation, and
don't change the clip rect.

bug: 26435476
Change-Id: Ic1ac71bfc814afd62572027fa11663c80d903a2d