Visibility changes to allow access by SetupWizard
Also adding a method that is called when the SIM info changes,
to update sugestions.
This is a must for SetupWizard, where the language selection can
be displayed before the SIM is fully initialized.
Might also use it to update suggested locals when SIM status changes
(SIM card replaced, or the user crosses border to a different country)

Bug: 25797548
Change-Id: I1a68b4e71046caa3b2810b37c74be5eeb00a255c