Allow activities without FLAG_ACTIVITY_NEW_TASK set to be home activities
Some callers can launch the home activity without setting
FLAG_ACTIVITY_NEW_TASK on the intent. When this occurred we were
putting a copy of the home activity in the application stack
which causes a few issues like the home activity showing up in
the recents list.
We now don't check the FLAG_ACTIVITY_NEW_TASK flag when determining
if the activity is a home activity is it meets all the other
requirements.

Bug: 26797267
Change-Id: I8176c6d9c27d26e9db6fa9f0f1a48e2e1388fb46