Fixed some issues with layer adjustment for special windows.
- Adjust layers of all windows in pinned stack.
The surface view of a pinned activity will be hidden during transition
animations because we were only adjusting the layer of the main app
window when we adjust special windows in WindowLayersController.
We now adjust the layers of all windows in the pinned stack.
- Also adjust layers for all windows in the docked stack for the same
reasons
- Adjust layers for replacing windows that was lost when the code
was moved to the separate class.

Bug: 26697719
Change-Id: I42fc3f9a51de9999157bcc7fee25a334b3e93ea1